Isolate chgcar & locpot from poscar.py
